Solve a scope problem in a list comprehension
In the following code, the variable F1 can't be used in the list
because of the problem written in
https://stackoverflow.com/questions/32894942/listcomp-unable-to-ac
cess-locals-defined-in-code-called-by-exec-if-nested-in-fun.

```
echo 1 | opy '{ print(" ".join([ str(x)*F1 for x in range(3) ])) }'
```

At the fix, local variables are added to the global variables
before the evaluation. In the case of the above example, F1 is
treated as a global variable. With this hack, the list
comprehension becomes valid.